fixed start_attempts flag, moved exceedance tracker to another file
